Reservation rate report excel changes
Take last 31 characters from unit name and use
it as excel sheet name. Result may be a bit silly
but this way the string will contain as much
information as it can. This is done because excel
sheet name has a 31 character limit.

Include Unit field 'street_address' in excel
because there can be multiple Unit objects
with the same name but are differentiated by address /
floor.